requirements.txt
langchain==0.2.*
langchain-community>=0.2.0
langchain-openai>=0.1.7
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.3
chromadb==0.5.*
sentence-transformers>=5.1.1
openai>=1.30.0
tiktoken
python-dotenv